[DDW-423] Moving ADA actions
