+++ source/renderer/app/actions/wallet-settings-actions.js
@@ -1,5 +1,5 @@
 // @flow
-import Action from '../lib/Action';
+import Action from './lib/Action';
 
 export type WalletExportToFileParams = {
   walletId: string,
